Mudança na estrutura de algumas funções e retorno da biblioteca de IA para futuras implementações.

+++ src/lib/convert.ts
@@ -10,6 +10,7 @@ export async function convertMp4ToMp3 (sourceType: 'buffer' | 'url',  video: Buf
 
         if(sourceType == 'buffer'){
             if(!Buffer.isBuffer(video)) throw new Error("O tipo selecionado da mídia foi Buffer mas a mídia não é um Buffer.")
+
             fs.writeFileSync(inputVideoPath, video)
         } else if (sourceType == 'url'){
             if(typeof video != 'string') throw new Error("O tipo selecionado da mídia foi URL mas a mídia não é uma URL.")
@@ -51,22 +52,16 @@ export async function convertVideoToThumbnail(sourceType : "file"|"buffer"|"url"
         const outputThumbnailPath = getTempPath('jpg')
 
         if(sourceType == "file"){
-            if(typeof video !== 'string'){
-                throw new Error('O tipo de operação está definido como FILE mas a mídia enviada não é um caminho de arquivo válido.')
-            }
+            if(typeof video !== 'string') throw new Error('O tipo de operação está definido como FILE mas a mídia enviada não é um caminho de arquivo válido.')
 
             inputPath = video
         } else if(sourceType == "buffer"){
-            if(!Buffer.isBuffer(video)) {
-                throw new Error('O tipo de operação está definido como BUFFER mas a mídia enviada não é um buffer válido.')
-            }
+            if(!Buffer.isBuffer(video)) throw new Error('O tipo de operação está definido como BUFFER mas a mídia enviada não é um buffer válido.')
 
             inputPath = getTempPath('mp4')
             fs.writeFileSync(inputPath, video)
         } else if(sourceType == "url"){
-            if(typeof video !== 'string') {
-                throw new Error('O tipo de operação está definido como URL mas a mídia enviada não é uma url válida.')
-            }
+            if(typeof video !== 'string') throw new Error('O tipo de operação está definido como URL mas a mídia enviada não é uma url válida.')
 
             const responseUrlBuffer = await axios.get(video,  { responseType: 'arraybuffer' }).catch(()=>{
                 throw new Error("Houve um erro ao fazer download da mídia para a thumbnail, tente novamente mais tarde.")
@@ -89,9 +84,7 @@ export async function convertVideoToThumbnail(sourceType : "file"|"buffer"|"url"
             throw new Error("Houve um erro ao conversão a mídia para thumbnail, tente novamente mais tarde.")
         })
 
-        if(sourceType != 'file' && inputPath) {
-            fs.unlinkSync(inputPath)
-        }
+        if(sourceType != 'file' && inputPath) fs.unlinkSync(inputPath)
 
         const thumbBase64 : Base64URLString = fs.readFileSync(outputThumbnailPath).toString('base64')
         fs.unlinkSync(outputThumbnailPath)
